# Improve performance of endpoints deduplication